requirements.txt
python-dotenv
openai
azure-identity
azure-monitor-query
pandas
